[v0.90][WP-14] Split long-lived agent responsibilities (Closes #2107)
